# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 'Laravel API Documentation'
  description: 'Helpdesk Ticket API - A production-style REST API for managing support tickets'
  version: 1.0.0
servers:
  -
    url: 'http://helpdesk-ticket-api.test'
tags:
  -
    name: Authentication
    description: "\nAPIs for managing authentication tokens"
  -
    name: 'Health Check'
    description: "\nAPI health check endpoint"
  -
    name: 'Ticket Comments'
    description: "\nAPIs for managing ticket comments"
  -
    name: Tickets
    description: "\nAPIs for managing support tickets"
  -
    name: Users
    description: "\nAPIs for managing users (admin/agent only)"
paths:
  /api/v1/auth/token:
    post:
      summary: 'Create authentication token'
      operationId: createAuthenticationToken
      description: 'Generate a new API token for the user.'
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  status: success
                  message: 'Token created successfully.'
                  data:
                    token: 1|abc123...
                    abilities:
                      - 'tickets:view'
                      - 'tickets:create'
                properties:
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: 'Token created successfully.'
                  data:
                    type: object
                    properties:
                      token:
                        type: string
                        example: 1|abc123...
                      abilities:
                        type: array
                        example:
                          - 'tickets:view'
                          - 'tickets:create'
                        items:
                          type: string
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  description: "The user's email address."
                  example: admin@example.com
                password:
                  type: string
                  description: "The user's password."
                  example: password
                device_name:
                  type: string
                  description: 'The device name for the token.'
                  example: Postman
              required:
                - email
                - password
                - device_name
      security: []
    delete:
      summary: 'Revoke authentication token'
      operationId: revokeAuthenticationToken
      description: "Delete the current user's API token."
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  status: success
                  message: 'Token revoked successfully.'
                properties:
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: 'Token revoked successfully.'
      tags:
        - Authentication
  /api/v1/health:
    get:
      summary: 'Health check'
      operationId: healthCheck
      description: 'Check if the API is running.'
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  status: ok
                  version: v1
                properties:
                  status:
                    type: string
                    example: ok
                  version:
                    type: string
                    example: v1
      tags:
        - 'Health Check'
      security: []
  '/api/v1/tickets/{ticket_id}/comments':
    get:
      summary: 'List ticket comments'
      operationId: listTicketComments
      description: "Get all comments for a specific ticket.\nCustomers can only see public comments."
      parameters: []
      responses:
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  status: error
                  message: Unauthenticated.
                properties:
                  status:
                    type: string
                    example: error
                  message:
                    type: string
                    example: Unauthenticated.
      tags:
        - 'Ticket Comments'
    post:
      summary: 'Create ticket comment'
      operationId: createTicketComment
      description: "Add a new comment to a ticket.\nCustomers can only create public comments."
      parameters: []
      responses: {  }
      tags:
        - 'Ticket Comments'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                body:
                  type: string
                  description: 'Comment text (3-2000 characters).'
                  example: 'This issue has been resolved.'
                is_internal:
                  type: boolean
                  description: 'Whether the comment is internal (admin/agent only).'
                  example: false
              required:
                - body
    parameters:
      -
        in: path
        name: ticket_id
        description: 'The ID of the ticket.'
        example: 1
        required: true
        schema:
          type: integer
      -
        in: path
        name: ticket
        description: 'The ticket ID.'
        example: 1
        required: true
        schema:
          type: integer
  /api/v1/tickets:
    get:
      summary: 'List tickets'
      operationId: listTickets
      description: 'Get a paginated list of tickets with optional filtering and sorting.'
      parameters:
        -
          in: query
          name: include
          description: 'Comma-separated list of relationships to include (customer,assignedAgent,comments).'
          example: 'customer,comments'
          required: false
          schema:
            type: string
            description: 'Comma-separated list of relationships to include (customer,assignedAgent,comments).'
            example: 'customer,comments'
        -
          in: query
          name: 'filter[status]'
          description: 'Filter by status.'
          example: open
          required: false
          schema:
            type: string
            description: 'Filter by status.'
            example: open
        -
          in: query
          name: 'filter[priority]'
          description: 'Filter by priority.'
          example: high
          required: false
          schema:
            type: string
            description: 'Filter by priority.'
            example: high
        -
          in: query
          name: 'filter[customer_id]'
          description: 'Filter by customer ID.'
          example: 5
          required: false
          schema:
            type: integer
            description: 'Filter by customer ID.'
            example: 5
        -
          in: query
          name: 'filter[assigned_to]'
          description: 'Filter by assigned agent ID.'
          example: 2
          required: false
          schema:
            type: integer
            description: 'Filter by assigned agent ID.'
            example: 2
        -
          in: query
          name: 'filter[created_after]'
          description: 'Filter by creation date.'
          example: '2026-01-01'
          required: false
          schema:
            type: string
            description: 'Filter by creation date.'
            example: '2026-01-01'
        -
          in: query
          name: sort
          description: 'Sort by field(s). Prefix with - for descending.'
          example: '-created_at'
          required: false
          schema:
            type: string
            description: 'Sort by field(s). Prefix with - for descending.'
            example: '-created_at'
        -
          in: query
          name: page
          description: 'Page number.'
          example: 1
          required: false
          schema:
            type: integer
            description: 'Page number.'
            example: 1
        -
          in: query
          name: per_page
          description: 'Items per page.'
          example: 15
          required: false
          schema:
            type: integer
            description: 'Items per page.'
            example: 15
      responses:
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  status: error
                  message: Unauthenticated.
                properties:
                  status:
                    type: string
                    example: error
                  message:
                    type: string
                    example: Unauthenticated.
      tags:
        - Tickets
    post:
      summary: 'Create ticket'
      operationId: createTicket
      description: 'Create a new support ticket.'
      parameters: []
      responses: {  }
      tags:
        - Tickets
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  description: 'Ticket title (5-120 characters).'
                  example: 'Payment failed'
                description:
                  type: string
                  description: 'Ticket description (min 20 characters).'
                  example: 'I paid for the plan, but my account is not upgraded.'
                priority:
                  type: string
                  description: 'Priority level.'
                  example: high
                user_id:
                  type: integer
                  description: 'User ID (admin only).'
                  example: 5
              required:
                - title
                - description
                - priority
  '/api/v1/tickets/{id}':
    get:
      summary: 'Show ticket'
      operationId: showTicket
      description: 'Get details of a specific ticket.'
      parameters:
        -
          in: query
          name: include
          description: 'Comma-separated list of relationships to include.'
          example: 'customer,comments'
          required: false
          schema:
            type: string
            description: 'Comma-separated list of relationships to include.'
            example: 'customer,comments'
      responses:
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  status: error
                  message: Unauthenticated.
                properties:
                  status:
                    type: string
                    example: error
                  message:
                    type: string
                    example: Unauthenticated.
      tags:
        - Tickets
    put:
      summary: 'Replace ticket (PUT)'
      operationId: replaceTicketPUT
      description: 'Fully replace a ticket. All fields are required.'
      parameters: []
      responses: {  }
      tags:
        - Tickets
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  description: 'Ticket title (5-120 characters).'
                  example: 'Payment failed'
                description:
                  type: string
                  description: 'Ticket description (min 20 characters).'
                  example: 'Full description'
                status:
                  type: string
                  description: Status.
                  example: open
                priority:
                  type: string
                  description: 'Priority level.'
                  example: high
                assigned_to:
                  type: integer
                  description: 'Assigned agent ID.'
                  example: 2
                  nullable: true
              required:
                - title
                - description
                - status
                - priority
    delete:
      summary: 'Delete ticket'
      operationId: deleteTicket
      description: 'Delete a ticket. Customers can only delete their own open tickets.'
      parameters: []
      responses: {  }
      tags:
        - Tickets
    parameters:
      -
        in: path
        name: id
        description: 'The ID of the ticket.'
        example: 1
        required: true
        schema:
          type: integer
      -
        in: path
        name: ticket
        description: 'The ticket ID.'
        example: 1
        required: true
        schema:
          type: integer
  '/api/v1/tickets/{ticket_id}':
    patch:
      summary: 'Update ticket (PATCH)'
      operationId: updateTicketPATCH
      description: 'Partially update a ticket. Only provided fields will be updated.'
      parameters: []
      responses: {  }
      tags:
        - Tickets
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  description: 'Ticket title (5-120 characters).'
                  example: 'Payment issue resolved'
                description:
                  type: string
                  description: 'Ticket description (min 20 characters).'
                  example: 'Updated description'
                status:
                  type: string
                  description: 'Status (admin/agent only).'
                  example: in_progress
                priority:
                  type: string
                  description: 'Priority (admin/agent only).'
                  example: medium
                assigned_to:
                  type: integer
                  description: 'Assigned agent ID (admin/agent only).'
                  example: 2
    parameters:
      -
        in: path
        name: ticket_id
        description: 'The ID of the ticket.'
        example: 1
        required: true
        schema:
          type: integer
      -
        in: path
        name: ticket
        description: 'The ticket ID.'
        example: 1
        required: true
        schema:
          type: integer
  /api/v1/users:
    get:
      summary: 'List users'
      operationId: listUsers
      description: 'Get a list of all users. Only accessible by admin and agents.'
      parameters: []
      responses:
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  status: error
                  message: Unauthenticated.
                properties:
                  status:
                    type: string
                    example: error
                  message:
                    type: string
                    example: Unauthenticated.
      tags:
        - Users
  '/api/v1/users/{id}':
    get:
      summary: 'Show user'
      operationId: showUser
      description: 'Get details of a specific user. Only accessible by admin and agents.'
      parameters: []
      responses:
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  status: error
                  message: Unauthenticated.
                properties:
                  status:
                    type: string
                    example: error
                  message:
                    type: string
                    example: Unauthenticated.
      tags:
        - Users
    parameters:
      -
        in: path
        name: id
        description: 'The ID of the user.'
        example: 1
        required: true
        schema:
          type: integer
      -
        in: path
        name: user
        description: 'The user ID.'
        example: 1
        required: true
        schema:
          type: integer